Add destroyer capsules
Add a heap of options to configure just about everything

Fix a multiplayer incompatibility when a player joins after a nuke/poison got triggered and a player joined within the cooldown and a tree explodes. That could cause a desync due to the new player getting a different result (Fixes #1)

Change a few settings to minutes / seconds instead of ticks. Closes #2, Closes #3

Allow specifying chances as double